feat: added xss check for company related pages

+++ app/Http/Controllers/Client/CompanyController.php
@@ -8,6 +8,7 @@ use App\Http\Requests\Client\Company\EditCompanyFormRequest;
 use App\Models\Company;
 use App\Models\Requirement;
 use Illuminate\Http\Request;
+use ProtoneMedia\LaravelXssProtection\Middleware\XssCleanInput;
 
 class CompanyController extends Controller
 {
@@ -15,32 +16,32 @@ class CompanyController extends Controller
     public function __construct()
     {
         $this->middleware('client.validate.companies.max_count')->only(['create', 'store']);
+        $this->middleware(XssCleanInput::class)->only(['store', 'update']);
     }
 
     public function index()
     {
         $companies = auth('client')->user()->companies;
-        return view('client.companies.index')->with(compact('companies'));   
+        return view('client.companies.index')->with(compact('companies'));
     }
 
 
     public function show(Company $company)
-    {  
+    {
         $client_submitted_requirements = $company->requirements->pluck('id')->toArray();
         $missing_requirements = Requirement::whereNotIn('id', $client_submitted_requirements)->get();
-        $featured_projects = $company->projects()->orderBy('id', 'desc')->take(5)->get();        
+        $featured_projects = $company->projects()->orderBy('id', 'desc')->take(5)->get();
         return view('client.companies.show')->with(compact('company', 'client_submitted_requirements', 'missing_requirements', 'featured_projects'));
     }
 
     public function create()
     {
-        return view('client.companies.create');   
+        return view('client.companies.create');
     }
 
     public function edit(Company $company)
     {
-        if(!$company->checkIfUserOwned())
-        {
+        if (!$company->checkIfUserOwned()) {
             return redirect(route('client.companies.index'))->withErrors(['message' => 'Unauthorized Access!']);
         }
 
@@ -49,16 +50,13 @@ class CompanyController extends Controller
 
     public function store(CreateCompanyFormRequest $request)
     {
-        try 
-        {
+        try {
             // Company::create($request->all());
             auth('client')->user()->companies()->create($request->validated());
 
             return redirect(route('client.companies.index'))
-                ->with('success', 'Company was successfully created.');  
-        }
-        catch(\Exception $e)
-        {
+                ->with('success', 'Company was successfully created.');
+        } catch (\Exception $e) {
             return redirect()->back()->withErrors([
                 'Operation Failed!' => $e->getMessage()
             ]);
@@ -67,14 +65,11 @@ class CompanyController extends Controller
 
     public function update(EditCompanyFormRequest $request, Company $company)
     {
-        try 
-        {
-            $company->update($request->except('name'));
+        try {
+            $company->update($request->validated());
             return redirect(route('client.companies.index'))
-                    ->with('success', 'Company was successfully updated.');  
-        }
-        catch(\Exception $e)
-        {
+                ->with('success', 'Company was successfully updated.');
+        } catch (\Exception $e) {
             return redirect()->back()->withErrors([
                 'Operation Failed!' => $e->getMessage()
             ]);
@@ -83,14 +78,14 @@ class CompanyController extends Controller
 
     public function destroy(Company $company)
     {
-        try
-        {
+        try {
+            if (!$company->checkIfUserOwned()) {
+                return redirect(route('client.companies.index'))->withErrors(['message' => 'Unauthorized Access!']);
+            }
+
             $company->delete();
-            return redirect(route('client.companies.index'))
-            ->with('success', 'Company was successfully deleted.');  
-        }
-        catch(\Exception $e)
-        {
+            return redirect(route('client.companies.index'))->with('success', 'Company was successfully deleted.');
+        } catch (\Exception $e) {
             return redirect(route('client.companies.index'))->withErrors([
                 'Operation Failed!' => $e->getMessage()
             ]);
